Set org store link back to original mapping
I had set that shortcut to "dictionary lookup," which I never use, but
upon removing it I discovered that the key didn't seem to be bound to
this org function, though the docs say it should be. So I mapped it
myself.

Now you can use C-c l to store a link and C-c C-l to spit it out.