Redefine active and stuck projects
This patch redefines a stuck project as a top-level headline with no
TODO keywords that contains no "active" headlines.

Unfortunately, that means that a project task marked DONE that has an
"active" tag would make the project appear not-stuck (which is untrue;
there is no active task because the task marked active is also marked
DONE).

Org doesn't offer any way to implement more complex stuck item logic,
so instead I added a hook on state change that ensures that the
"active" tag is removed from any entry that changes to the DONE
state. By definition, a task cannot be active *and* DONE.

This patch also renames "Stuck Projects" to "Waiting Projects," which
feels more accurate, and uses the "active" tag to divide tasks between
the main agenda and the backlog/review view.